% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banrep_cache}
\alias{banrep_cache}
\title{Catálogo interno de indicadores disponibles}
\format{
A data.table with five columns
\enumerate{
\item \code{idGrupo}: ID of the parent menu group
\item \code{NombreGrupo}: Name of the parent menu group
\item \code{idSerie}: ID of the indicator
\item \code{NombreSerie}: Name of the indicator
\item \code{idNombreSerie}: Category code of the indicator (used by \code{banrep_data()})
}
}
\source{
\url{https://suameca.banrep.gov.co/estadisticas-economicas/#/}
}
\usage{
banrep_cache
}
\description{
A snapshot of \code{catalogo_banrep()}, bundled with the package as an offline
fallback so \code{banrep_data()} does not need to hit the network just to
validate/look up an indicator code.
}
\examples{
banrep_cache

}
\keyword{datasets}
