% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{series_plot}
\alias{series_plot}
\title{Plot a Banrep series}
\usage{
series_plot(x, ...)
}
\arguments{
\item{x}{A data.table returned by \code{fetch_banrep()}/\code{banrep_data()}.}

\item{...}{Further arguments passed to methods.}
}
\value{
A ggplot object.
}
\description{
Plot a Banrep series
}
